V2 push dashboard

+++ AngularAuthUI/src/app/components/book-list/book-list.component.ts
@@ -10,6 +10,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { BookAddEditComponent } from '../book-add-edit/book-add-edit.component';
 import { CoreService } from '../../services/core.service';
 import { Router } from '@angular/router';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-book-list', 
@@ -42,8 +43,9 @@ export class BookListComponent implements OnInit {
     private bookService: BookService,
     private fb: FormBuilder,
     private _dialog: MatDialog,
+    private dialog: MatDialog,
     private _coreService: CoreService,
-    private router: Router 
+    private router: Router,
 
   ) {
     this.bookForm = this.fb.group({
@@ -123,13 +125,25 @@ export class BookListComponent implements OnInit {
   }
 
   deleteBook(id: number) {
-    this.bookService.deleteBook(id).subscribe({
-      next: (res) => {
-        this._coreService.openSnackBar('Book deleted!', 'done')
-        this.loadBooks();
-      },
-      error: console.log
-    })
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '300px',
+      
+    });   
+  
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) { 
+        this.bookService.deleteBook(id).subscribe({
+          next: () => {
+            this._coreService.openSnackBar('Book deleted!', 'done');
+            this.loadBooks();
+          },
+          error: (err) => {
+            console.error('Error deleting book:', err);
+            this._coreService.openSnackBar('Error deleting book. Please try again.');
+          }
+        });
+      }
+    });
   }
 
   openAddEditBookForm() {
